Move AccountLocks to separate module
